allow for multiple inputs
