+++ package.json
@@ -11,7 +11,6 @@
   ],
   "scripts": {
     "clean": "rimraf .upload && rimraf packages/*/dist",
-    "postinstall": "patch-package",
     "lint": "eslint --ext .ts,.tsx .",
     "lint:fix": "eslint --ext .ts,.tsx . --fix",
     "start": "npm run --ws start --if-present",
@@ -44,6 +43,7 @@
     "@types/react-color": "^3.0.12",
     "@types/react-dom": "^18.3.0",
     "@types/react-syntax-highlighter": "^15.5.13",
+    "@types/semver": "^7.5.8",
     "@typescript-eslint/eslint-plugin": "^7.12.0",
     "@typescript-eslint/parser": "^7.12.0",
     "@vitejs/plugin-react": "^4.3.0",
@@ -51,21 +51,21 @@
     "autoprefixer": "^10.4.19",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.2",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
     "husky": "^9.0.11",
     "inquirer": "^9.2.23",
     "npm-run-all": "^4.1.5",
-    "patch-package": "^8.0.0",
     "postcss": "^8.4.38",
     "pre-commit": "^1.2.2",
     "prettier": "^3.3.0",
     "prettier-plugin-tailwindcss": "^0.6.1",
     "rimraf": "^5.0.7",
+    "semver": "^7.6.2",
     "tailwindcss": "^3.4.3",
     "tsx": "^4.11.2",
     "typescript": "^5.4.5",
