+++ packages/frontend/src/components/properties/color-property.tsx
@@ -1,0 +1,7 @@
+import type { ColorProps } from '../ui/color';
+import { ColorInput } from '../ui/color';
+import type { PropertyPanelData } from './propertyTypes';
+
+export const ColorProperty: React.FC<PropertyPanelData> = ({ value, entry }) => {
+  return <ColorInput {...(entry.options as ColorProps)} value={value} onChange={entry.onChange} />;
+};
