+++ packages/frontend/src/pages/scene/scene-section/Properties.tsx
@@ -1,32 +1,15 @@
-import type { SliderProps } from '@radix-ui/react-slider';
-import type { SwitchProps } from '@radix-ui/react-switch';
 import Fuse from 'fuse.js';
-import { memo, useCallback, useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { FaCopy as CopyIcon } from 'react-icons/fa6';
 import { useDevtoolStore } from '../../../App';
 import { CollapsibleSection } from '../../../components/collapsible/collapsible-section';
-import type { ColorProps } from '../../../components/properties/color';
-import { ColorInput } from '../../../components/properties/color';
-import { PropertyEntry } from '../../../components/properties/property';
-import type { Vector2Props, VectorXProps } from '../../../components/properties/vector2';
-import { Vector2, VectorX } from '../../../components/properties/vector2';
-import type { ButtonProps } from '../../../components/ui/button';
+import { PropertyEntry } from '../../../components/properties/propertyEntry';
+import type { PropertyPanelData } from '../../../components/properties/propertyTypes';
+import { propertyMap } from '../../../components/properties/propertyTypes';
 import { Button } from '../../../components/ui/button';
-import type { InputProps } from '../../../components/ui/input';
-import { Input } from '../../../components/ui/input';
-import {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '../../../components/ui/select';
 import { Separator } from '../../../components/ui/separator';
-import { Slider } from '../../../components/ui/slider';
-import { Switch } from '../../../components/ui/switch';
 import { TooltipWrapper } from '../../../components/ui/tooltip';
-import { copyToClipboard, formatCamelCase, isDifferent } from '../../../lib/utils';
+import { copyToClipboard, formatCamelCase } from '../../../lib/utils';
 
 interface PanelProps {
   children: React.ReactNode;
@@ -84,137 +67,6 @@ const Panel: React.FC<PanelProps> = ({ children, onSearch, onCopy }) => {
       </div>
     </>
   );
-};
-
-export type PropertyTypes =
-  | 'boolean'
-  | 'number'
-  | 'range'
-  | 'select'
-  | 'text'
-  | 'button'
-  | 'vector2'
-  | 'vectorX'
-  | 'color';
-type SelectionProps = { options: string[] };
-type ButtonFnProps = { label: string } & ButtonProps;
-export type PropertyPanelData = {
-  value: any;
-  prop: string;
-  entry: {
-    section: string;
-    tooltip?: string;
-    label?: string;
-    type: PropertyTypes;
-    options?: InputProps | SwitchProps | Vector2Props | SliderProps | SelectionProps | any;
-    onChange: (value: string | number | boolean) => void;
-  };
-};
-
-const BooleanProperty: React.FC<PropertyPanelData> = ({ value, entry }) => {
-  return (
-    <Switch
-      className="bg-border h-6"
-      {...(entry.options as SwitchProps)}
-      checked={value}
-      onCheckedChange={entry.onChange}
-    />
-  );
-};
-
-const NumberProperty: React.FC<PropertyPanelData> = ({ value, entry }) => {
-  return (
-    <Input
-      {...entry.options}
-      type="number"
-      value={value}
-      onChange={(e) => entry.onChange(Number(e.target.value))}
-      className="border-border hover:border-secondary focus:border-secondary h-6 w-full rounded text-xs outline-none"
-    />
-  );
-};
-
-const Vector2Property: React.FC<PropertyPanelData> = ({ value, entry }) => {
-  return <Vector2 {...(entry.options as Vector2Props)} onChange={entry.onChange} value={value} />;
-};
-
-const RangeProperty: React.FC<PropertyPanelData> = ({ value, entry }) => {
-  return (
-    <Slider
-      {...(entry.options as SliderProps)}
-      value={[value]}
-      onValueChange={(value) => entry.onChange(JSON.stringify(value))}
-      className="border-border hover:border-secondary focus:border-secondary h-6 w-full rounded outline-none"
-    />
-  );
-};
-
-const SelectProperty: React.FC<PropertyPanelData> = ({ value, entry }) => {
-  return (
-    <Select value={value} onValueChange={(value) => entry.onChange(JSON.stringify(value))}>
-      <SelectTrigger className="border-border hover:border-secondary focus:border-secondary h-7 w-full text-sm outline-none">
-        <SelectValue />
-      </SelectTrigger>
-      <SelectContent>
-        <SelectGroup>
-          {(entry.options as SelectionProps).options.map((option: string) => (
-            <SelectItem key={option} value={option}>
-              {option}
-            </SelectItem>
-          ))}
-        </SelectGroup>
-      </SelectContent>
-    </Select>
-  );
-};
-
-const ButtonProperty: React.FC<PropertyPanelData> = ({ entry }) => {
-  return (
-    <Button
-      variant="outline"
-      size="sm"
-      className="border-border hover:border-secondary focus:border-secondary h-6 w-full rounded outline-none"
-      onClick={() => entry.onChange(true)}
-      {...(entry.options as ButtonFnProps)}
-    >
-      {(entry.options as ButtonFnProps)?.label ?? entry.label ?? 'Click'}
-    </Button>
-  );
-};
-
-const VectorXProperty: React.FC<PropertyPanelData> = ({ value, entry }) => {
-  return <VectorX {...(entry.options as VectorXProps)} onChange={entry.onChange} value={value} />;
-};
-
-const ColorProperty: React.FC<PropertyPanelData> = ({ value, entry }) => {
-  return <ColorInput {...(entry.options as ColorProps)} value={value} onChange={entry.onChange} />;
-};
-
-const TextProperty: React.FC<PropertyPanelData> = ({ value, entry }) => {
-  return (
-    <Input
-      {...entry.options}
-      type="text"
-      value={value}
-      onChange={(e) => entry.onChange(JSON.stringify(e.target.value))}
-      className="border-border hover:border-secondary focus:border-secondary h-6 w-full rounded text-xs outline-none"
-    />
-  );
-};
-
-const memoTest = (props: any, nextProps: any) => {
-  return !isDifferent(props, nextProps);
-};
-const properties: Record<PropertyTypes, React.FC<PropertyPanelData>> = {
-  boolean: memo(BooleanProperty, memoTest),
-  number: memo(NumberProperty, memoTest),
-  vector2: memo(Vector2Property, memoTest),
-  range: memo(RangeProperty, memoTest),
-  select: memo(SelectProperty, memoTest),
-  text: memo(TextProperty, memoTest),
-  button: memo(ButtonProperty, memoTest),
-  vectorX: memo(VectorXProperty, memoTest),
-  color: memo(ColorProperty, memoTest),
 };
 
 function clone(obj: any) {
@@ -302,7 +154,7 @@ export const Properties: React.FC = () => {
         <CollapsibleSection key={section} title={section} className="border-x" onCopy={() => onCopy(section)}>
           <div className="p-2 [&>*:first-child]:pt-0">
             {sections[section].map((prop, i) => {
-              const Component = properties[prop.entry.type];
+              const Component = propertyMap[prop.entry.type];
               return (
                 <PropertyEntry
                   key={`${prop.prop}-${i}`}
