+++ packages/frontend/src/pages/scene/state.ts
@@ -1,5 +1,5 @@
 import type { ZustSet } from '../../lib/utils';
-import type { PropertyPanelData } from './scene-section/Properties';
+import type { PropertyPanelData } from '../../components/properties/propertyTypes';
 
 export interface SceneState {
   stats: Record<string, number> | null;
